backend/fastapi/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
autopep8==2.3.2
certifi==2025.1.31
click==8.1.8
colorama==0.4.6
Django==5.1.5
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-jwt==1.11.0
djangorestframework_simplejwt==5.4.0
dnspython==2.7.0
drf-yasg==1.21.8
email_validator==2.2.0
fastapi==0.104.1
fastapi-cli==0.0.7
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.5
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mysqlclient==2.2.7
orjson==3.10.15
packaging==24.2
pycodestyle==2.12.1
pydantic==2.10.6
pydantic-extra-types==2.10.2
pydantic-settings==2.7.1
pydantic_core==2.27.2
Pygments==2.19.1
PyJWT==1.7.1
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2025.1
PyYAML==6.0.2
rich==13.9.4
rich-toolkit==0.13.2
shellingham==1.5.4
sniffio==1.3.1
sqlparse==0.5.3
starlette==0.45.3
toml==0.10.2
typer==0.15.1
typing_extensions==4.12.2
tzdata==2025.1
ujson==5.10.0
uritemplate==4.1.1
uvicorn==0.24.0
watchfiles==1.0.4
websockets==14.2
paddleocr==2.7.0
pillow==10.1.0
numpy==1.26.2
pandas==2.1.3
scikit-learn==1.3.2
mysql-connector-python==8.2.0
openai==1.3.5
